add check for categories still in use on delete

+++ ReadLater5/ReadLater5/Controllers/CategoriesController.cs
@@ -108,8 +108,15 @@ namespace ReadLater5.Controllers
         public IActionResult DeleteConfirmed(int id)
         {
             Category category = _categoryService.GetCategory(id);
-            _categoryService.DeleteCategory(category);
-            return RedirectToAction("Index");
+            if (_categoryService.IsCategoryUsed(category))
+            {
+                return new StatusCodeResult(Microsoft.AspNetCore.Http.StatusCodes.Status409Conflict);
+            }
+            else
+            {
+                _categoryService.DeleteCategory(category);
+                return RedirectToAction("Index");
+            }
         }
     }
 }
